+++ part2/phonebook/src/App.jsx
@@ -1,61 +1,9 @@
 import { useEffect, useState } from "react";
 import "./App.css";
 import axios from "axios";
-
-function FilterShown({persons}) {
-	const [filter, setfilter] = useState("");
-
-	const searchChange = (event) => {
-		const filtrar = event.target.value.toLowerCase();
-		setfilter(filtrar);
-	};
-
-	return (
-		<section>
-			<div>
-				<label htmlFor="search-input">filter: </label>
-				<input id="search-input" type="text" onChange={searchChange} name="search"/>
-			</div>
-
-			{persons.map((value) => {
-				if (value.name.toLowerCase().includes(filter)) {		
-					return <h3 key={value.id}>{value.name}, {value.number} </h3>;
-				}
-			})}
-		</section>
-	);
-}
-function Form({ handleChange, handleSubmit }) {
-	return (
-		<section>
-			<h2>add a new</h2>
-			<form onSubmit={handleSubmit}>
-				<div>
-					<label htmlFor="name-input">Name: </label>
-					<input id="name-input" type="text" onChange={handleChange} name="name"/>
-				</div>
-				<div>
-					<label htmlFor="number-input">Number: </label>
-					<input id="number-input" type="number" onChange={handleChange} name="number"/>
-				</div>
-				<div>
-					<button type="submit">add</button>
-				</div>
-			</form>
-		</section>
-	);
-}
-
-function PersonsList({ persons}) {
-	return (
-		<section>
-			<h2>Numbers</h2>
-			{persons.map((value) => {
-				return <h3 key={value.id}>{value.name}, {value.number} </h3>;
-			})}
-		</section>
-	);
-}
+import { FilterShown } from "./components/FilterShown.jsx";
+import { Form } from "./components/Form.jsx";
+import { PersonsList } from "./components/PersonsList.jsx";
 
 function App() {
 	const [persons, setPersons] = useState([]);
@@ -70,7 +18,6 @@ function App() {
 	}, []);
 
 	const handleChange = (event) => {
-		console.log(event);
 		const { name, value } = event.target;
 		if (name === "name") {
 			setNewName(value);
@@ -90,10 +37,14 @@ function App() {
 		const personToAddToState = {
 			name: newName,
 			number: newNumber,
-			id: persons.length + 1,
 		};
 
-		setPersons(persons.concat(personToAddToState));
+		axios.post("http://localhost:3001/persons", personToAddToState).then((response) => {
+			const {data} = response;
+			
+			setPersons((prevPersons) => prevPersons.concat(data));
+		});
+
 	};
 
 	return (
